test: refactored tests for LikesController

+++ app/controllers/likes_controller.rb
@@ -2,14 +2,16 @@ class LikesController < ApplicationController
 
   def create
     post = Post.find(params[:post_id])
-    liked = Like.find_by(post_id: post.id, user_id: current_user.id)
+    if !!current_user
+      liked = Like.find_by(post_id: post.id, user_id: current_user.id)
 
-    if !liked
-      like = Like.create(post_id: post.id, user_id: current_user.id)
-    else
-      liked.destroy
+      if !liked
+        like = Like.create(post_id: post.id, user_id: current_user.id)
+      else
+        liked.destroy
+      end
+      
     end
-
     redirect_to posts_path
   end
